perf(group): fixed slim GroupHero + memoized feed build
Extract the group hero into a fixed <GroupHero/> that no longer scrolls
with the list, and shrink it (title-size money, tighter gradient, 42px
action circles). The pending-confirmation deck lives in the hero: inline
confirm/reject for a single claim, a summary slide routing to the
dedicated review screen for two or more.

Wrap the feed build (visibleExpenses, expenseSections, feedItems,
balanceItems, activityItems) in useMemo so a re-render that leaves the
ledger untouched no longer re-sections every expense — the tab switch and
navigation no longer recompute the whole list.
